+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessCSSFilesTask.java
@@ -56,6 +56,7 @@ public class ProcessCSSFilesTask extends ProcessFilesTask {
      * @param outputDir directory to write the final file
      * @param outputFilename the output file name
      * @param suffix final filename suffix
+     * @param nosuffix whether to use a suffix for the minified filename or not
      * @param charset if a character set is specified, a byte-to-char variant allows the encoding to be selected.
      *        Otherwise, only byte-to-byte operations are used
      * @param linebreak split long lines after a specific column
@@ -63,9 +64,9 @@ public class ProcessCSSFilesTask extends ProcessFilesTask {
     public ProcessCSSFilesTask(Log log, Integer bufferSize, boolean debug, boolean skipMerge, boolean skipMinify,
             String webappSourceDir, String webappTargetDir, String inputDir, List<String> sourceFiles,
             List<String> sourceIncludes, List<String> sourceExcludes, String outputDir, String outputFilename,
-            String suffix, String charset, int linebreak) {
+            String suffix, boolean nosuffix, String charset, int linebreak) {
         super(log, bufferSize, debug, skipMerge, skipMinify, webappSourceDir, webappTargetDir, inputDir, sourceFiles,
-                sourceIncludes, sourceExcludes, outputDir, outputFilename, suffix, charset, linebreak);
+                sourceIncludes, sourceExcludes, outputDir, outputFilename, suffix, nosuffix, charset, linebreak);
     }
 
     /**
@@ -82,8 +83,12 @@ public class ProcessCSSFilesTask extends ProcessFilesTask {
                     OutputStream out = new FileOutputStream(minifiedFile);
                     InputStreamReader reader = new InputStreamReader(in, charset);
                     OutputStreamWriter writer = new OutputStreamWriter(out, charset)) {
-                log.info("Creating minified file [" + ((debug) ? minifiedFile.getPath() : minifiedFile.getName())
-                        + "].");
+                if (debug) {
+                    log.info("Creating minified file [" + minifiedFile.getPath() + "].");
+                } else {
+                    File temp = (nosuffix) ? mergedFile : minifiedFile;
+                    log.info("Creating minified file [" + temp.getName() + "].");
+                }
 
                 CssCompressor compressor = new CssCompressor(reader);
                 compressor.compress(writer, linebreak);
@@ -91,6 +96,8 @@ public class ProcessCSSFilesTask extends ProcessFilesTask {
                 log.error(e.getMessage(), e);
                 throw e;
             }
+
+            cleanupFiles(mergedFile, minifiedFile);
         }
     }
 }