+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessFilesTask.java
@@ -28,6 +28,7 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.io.SequenceInputStream;
+import java.net.URI;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -56,13 +57,15 @@ public abstract class ProcessFilesTask implements Callable<Object> {
 
     protected boolean skipMinify;
 
-    private String mergedFilename;
-
-    private String suffix;
-
     protected String charset;
 
     protected int linebreak;
+
+    protected boolean nosuffix;
+
+    private String mergedFilename;
+
+    private String suffix;
 
     private File targetDir;
 
@@ -91,6 +94,7 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      * @param outputDir directory to write the final file
      * @param outputFilename the output file name
      * @param suffix final filename suffix
+     * @param nosuffix whether to use a suffix for the minified filename or not
      * @param charset if a character set is specified, a byte-to-char variant allows the encoding to be selected.
      *        Otherwise, only byte-to-byte operations are used
      * @param linebreak split long lines after a specific column
@@ -98,16 +102,17 @@ public abstract class ProcessFilesTask implements Callable<Object> {
     public ProcessFilesTask(Log log, Integer bufferSize, boolean debug, boolean skipMerge, boolean skipMinify,
             String webappSourceDir, String webappTargetDir, String inputDir, List<String> sourceFiles,
             List<String> sourceIncludes, List<String> sourceExcludes, String outputDir, String outputFilename,
-            String suffix, String charset, int linebreak) {
+            String suffix, boolean nosuffix, String charset, int linebreak) {
         this.log = log;
         this.bufferSize = bufferSize;
         this.debug = debug;
         this.skipMerge = skipMerge;
         this.skipMinify = skipMinify;
-        this.mergedFilename = outputFilename;
-        this.suffix = suffix;
         this.charset = charset;
         this.linebreak = linebreak;
+        this.nosuffix = nosuffix;
+        this.mergedFilename = outputFilename;
+        this.suffix = suffix;
 
         File sourceDir = new File(webappSourceDir + File.separator + inputDir);
         for (String sourceFilename : sourceFiles) {
@@ -145,8 +150,8 @@ public abstract class ProcessFilesTask implements Callable<Object> {
                 log.info("Skipping minify step.");
             } else {
                 File mergedFile = new File(targetDir, mergedFilename);
-                File minifiedFile = new File(targetDir, mergedFile.getName().replace(extension, suffix + extension));
                 merge(mergedFile);
+                File minifiedFile = new File(targetDir, mergedFilename.replace(extension, suffix + extension));
                 minify(mergedFile, minifiedFile);
             }
         } else if (!sourceFilesEmpty || !sourceIncludesEmpty) {
@@ -164,7 +169,7 @@ public abstract class ProcessFilesTask implements Callable<Object> {
      * @param mergedFile output file resulting from the merged step
      * @throws IOException when the merge step fails
      */
-    private void merge(File mergedFile) throws IOException {
+    protected void merge(File mergedFile) throws IOException {
         if (mergedFile != null) {
             try (InputStream sequence = new SequenceInputStream(new ListOfFiles(log, files, debug));
                     OutputStream out = new FileOutputStream(mergedFile);
@@ -183,11 +188,34 @@ public abstract class ProcessFilesTask implements Callable<Object> {
     /**
      * Minifies source file.
      *
-     * @param file input file resulting from the merged step
+     * @param mergedFile input file resulting from the merged step
      * @param minifiedFile output file resulting from the minify step
      * @throws IOException when the minify step fails
      */
-    abstract void minify(File file, File minifiedFile) throws IOException;
+    abstract void minify(File mergedFile, File minifiedFile) throws IOException;
+
+    /**
+     * Cleanup files. Remove merged file is nosuffix parameter was set to true.
+     *
+     * @param mergedFile input file resulting from the merged step
+     * @param minifiedFile output file resulting from the minify step
+     */
+    void cleanupFiles(File mergedFile, File minifiedFile) {
+        if (nosuffix) {
+            String mergedFilename = mergedFile.getName();
+            URI mergedFileURI = mergedFile.toURI();
+
+            if (debug) {
+                log.info("Deleting the file [" + mergedFilename + "].");
+            }
+            mergedFile.delete();
+
+            if (debug) {
+                log.info("Renaming the file [" + minifiedFile.getName() + "] to [" + mergedFilename + "].");
+            }
+            minifiedFile.renameTo(new File(mergedFileURI));
+        }
+    }
 
     /**
      * Logs an addition of a new source file.