+++ src/main/java/com/samaxes/maven/minify/plugin/ProcessJSFilesTask.java
@@ -74,6 +74,7 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
      * @param outputDir directory to write the final file
      * @param outputFilename the output file name
      * @param suffix final filename suffix
+     * @param nosuffix whether to use a suffix for the minified filename or not
      * @param charset if a character set is specified, a byte-to-char variant allows the encoding to be selected.
      *        Otherwise, only byte-to-byte operations are used
      * @param linebreak split long lines after a specific column
@@ -85,10 +86,10 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
     public ProcessJSFilesTask(Log log, Integer bufferSize, boolean debug, boolean skipMerge, boolean skipMinify,
             String webappSourceDir, String webappTargetDir, String inputDir, List<String> sourceFiles,
             List<String> sourceIncludes, List<String> sourceExcludes, String outputDir, String outputFilename,
-            String suffix, String charset, int linebreak, String jsEngine, boolean munge, boolean verbose,
-            boolean preserveAllSemiColons, boolean disableOptimizations) {
+            String suffix, boolean nosuffix, String charset, int linebreak, String jsEngine, boolean munge,
+            boolean verbose, boolean preserveAllSemiColons, boolean disableOptimizations) {
         super(log, bufferSize, debug, skipMerge, skipMinify, webappSourceDir, webappTargetDir, inputDir, sourceFiles,
-                sourceIncludes, sourceExcludes, outputDir, outputFilename, suffix, charset, linebreak);
+                sourceIncludes, sourceExcludes, outputDir, outputFilename, suffix, nosuffix, charset, linebreak);
 
         this.jsEngine = jsEngine;
         this.munge = munge;
@@ -111,8 +112,12 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
                     OutputStream out = new FileOutputStream(minifiedFile);
                     InputStreamReader reader = new InputStreamReader(in, charset);
                     OutputStreamWriter writer = new OutputStreamWriter(out, charset)) {
-                log.info("Creating minified file [" + ((debug) ? minifiedFile.getPath() : minifiedFile.getName())
-                        + "].");
+                if (debug) {
+                    log.info("Creating minified file [" + minifiedFile.getPath() + "].");
+                } else {
+                    File temp = (nosuffix) ? mergedFile : minifiedFile;
+                    log.info("Creating minified file [" + temp.getName() + "].");
+                }
 
                 if ("closure".equals(jsEngine)) {
                     log.debug("Using JavaScript compressor engine [Google Closure Compiler].");
@@ -139,6 +144,8 @@ public class ProcessJSFilesTask extends ProcessFilesTask {
                 log.error(e.getMessage(), e);
                 throw e;
             }
+
+            cleanupFiles(mergedFile, minifiedFile);
         }
     }
 }